refactor: use overmind data in issues table component
