+++ render/views/__tests__/AppViewPages/SummaryPage.spec.jsx
@@ -9,7 +9,7 @@ import MockAdapter from 'axios-mock-adapter';
 import { ThemeProvider } from 'styled-components';
 
 import theme from '../../../theme';
-import SummaryPage from '../../AppViewPages/SummaryPage';
+import { SummaryPage } from '../../AppViewPages/SummaryPage';
 import issueActions, { ISSUES_GET_PAGE } from '../../../actions/issues.actions';
 import { initialize, getInstance, reset } from '../../../../common/request';
